+++ ll/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.10.4] - 2022-07-14
 ### Added
 - Added a new W5500 type for VDM transfers with an infallible GPIO and an infallible SPI bus.
 - Added `defmt::Format` implementations for all register structs.
@@ -158,7 +158,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-01-02
 - Initial release
 
-[Unreleased]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.3...HEAD
+[Unreleased]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.4...HEAD
+[0.10.3]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.3...ll%2Fv0.10.4
 [0.10.3]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.2...ll%2Fv0.10.3
 [0.10.2]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.1...ll%2Fv0.10.2
 [0.10.1]: https://github.com/newAM/w5500-rs/compare/ll%2Fv0.10.0...ll%2Fv0.10.1